Fix rootfs source list handling in image.bbclass
Ensure that the APT sources list is only moved if both the sources-list
file and the sources.list.d directory exist. This prevents errors when
APT is not installed, as the sources.list.d directory will be missing.

+++ meta/classes/image.bbclass
@@ -419,7 +419,7 @@ do_rootfs_finalize() {
                 -maxdepth 1 -name 'qemu-*-static' -type f -delete
         fi
 
-        if [ -e "${ROOTFSDIR}/etc/apt/sources-list" ]; then
+        if [ -e "${ROOTFSDIR}/etc/apt/sources-list" ] && [ -e "${ROOTFSDIR}/etc/apt/sources-list.d" ]; then
             mv "${ROOTFSDIR}/etc/apt/sources-list" \
                 "${ROOTFSDIR}/etc/apt/sources.list.d/bootstrap.list"
         fi
